pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ds-utils"
version = "0.1.0"
description = "A Python utilities library for data science and machine learning, focusing on plotting, ML performance evaluation, and data cleaning"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Data Science Team" }
]
keywords = [
    "data-science",
    "machine-learning",
    "plotting",
    "visualization",
    "data-cleaning",
    "ml-evaluation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
]
dependencies = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "pyjanitor>=0.26",
    "pandera>=0.17",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.4",
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.23",
]

[project.urls]
Homepage = "https://github.com/your-org/ds-utils"
Documentation = "https://your-org.github.io/ds-utils"
Repository = "https://github.com/your-org/ds-utils"

[tool.hatch.build.targets.wheel]
packages = ["src/ds_utils"]

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "D",      # pydocstyle
]
ignore = [
    "D100",   # Missing docstring in public module
    "D104",   # Missing docstring in public package
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --cov=ds_utils --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["numpy.typing.mypy_plugin"]

[[tool.mypy.overrides]]
module = [
    "seaborn.*",
    "sklearn.*",
    "janitor.*",
    "pandera.*",
]
ignore_missing_imports = true
